Basic files for using cs-content (fast templating won't work, and it's old).
Set a constant that ensures site_config.php won't make any attempts to upgrade 
(since doing so would fail, anyway).  Created an empty library for the initial 
setup, and a minimal set of templates for cs-content...  I'm going to try to 
avoid converting everything to the cs-content system for a future release, but 
I guarantee nothing.