Testing the host & db existence works, creating the database works.
There's still a lack of good information when something goes wrong: if the 
database exists, the error displayed is very bad, as though it didn't set the 
host correctly.  But the database gets created, and I've confirmed that the 
schema files will load in their given order.

NOTE: setup should give the option for specifying an alternate user for normal 
access, whilst a special user (i.e. "postgres") can be used for setup.  It 
should be possible to create the specified "normal" user, as well... not sure 
if that's a per-db thing, or if a user is good for ALL the databases in the 
given installation... that's something for later, though.

/includes/setup/2.inc:
	* added a temporary class to handle creating the database & loading the 
	schema.  Presently, the database is created successfully if none exists... 
	the method that loads schema is only a stub right now.
/templates/system/message_box.tmpl:
	* required system file for contentSystem.