Semi-working update to handle the new config settings.
NOTE: the whole site will probably be converted to use the new cs_phpDB class, 
provided with cs-content, so the old phpDB class can (eventually) be removed 
entirely.  This commit doesn't make everything work by ANY MEANS: in fact, it 
only does enough to make the login page appear, though it shows an error at the 
bottom of the page.

/lib/genericPageClass.php:
	* initialize_locals():
		-- use cs_phpDB, and pass the output of get_config_db_params() to the 
		call to cs_phpDB::connect().  Woot.
/lib/globalFunctions.php:
	* get_config_db_params() [NEW]:
		-- confirms all required CONSTANTS are available for database connects, 
		and returns an array of parameters useful for cs_phpDB::connect().
/lib/logsClass.php:
	* __construct():
		-- requires cs_phpDB instead of phpDB.
/lib/upgradeClass.php:
	* MAIN::
		-- require the normal globalFunctions.php file, along with 
		cs-content/cs_phpDB.php.
	* get_database_version():
		-- requirements for the database abstraction layer are done at the top 
		of the library now.
		-- uses cs_phpDB, and passes the values from get_config_db_params() to 
		cs_phpDB::connect().