Move edit links to (mostly) match the view for projects.
NOTE: the "comment" box was changed into a hidden input, like the other stuff, 
so they have to click a link to display it.  This saves room, and should allow 
for the "priority" and "progress" items to be moved to a location that is more 
uniform to the rest of the page, and not hidden at the bottom.