Only automatically modify the status if it's not already set.
NOTE: status_id's should definitely be included in the config, or otherwise be 
dynamically pulled, so as to avoid hard-coding something that might break in the 
future. HINT: what if admins are allowed to create, edit, or even delete these 
records (maybe only a few of them, like "new", "running" and "closed" should be 
required, with an extra column in the DB to indicate that: all others can be 
intermediary and optional).

/lib/mainRecordClass.php:
	* perform_auto_updates():
		-- only modify status_id if it's not already set: previously, I would 
		attempt to change the status to "Running/Accepted", and it would get 
		reverted to "Pending".