Recursive cleaning by cleanString()...
Basically, send_email() was calling cleanString() to clean the array of notes. 
I believe this fix will not affect other things... but be warned.  cleanString() 
is used by a lot of very important things, such as string_from_array(), which is 
used to build the SQL queries... (*sweat*).

AFFECTS ISSUE:
#29: html_entity_decode() errors