Fix problem with version suffixes in setup & unit test versions.
Instead of having to update the unit tests every time an external library is 
updated (or having to make another release to fix the problem), now there's a 
function that will return the required external lib versions.  That function is 
the same as the one used by cs-project directly to ensure the versions it 
requires are available.

/includes/setup/3.inc:
	* __setupDefaultValues::set_version():
		-- append the suffix only if it has a length ("create_list()" seems to 
		append the "-" even if the suffix is zero-length).
/lib/globalFunctions.php:
	* get_required_external_lib_versions() [NEW]:
		-- retrieves the entire list of libs & versions, or just a single part 
		of that array (to make unit tests easier).
	* check_external_lib_versions():
		-- uses get_required_external_lib_versions() instead of arbitrarily 
		setting the array itself.
/lib/_unitTests_/cs-content_tests.php:
	* require globalFunctions.php
	* use the new get_required_external_lib_versions() to test the required 
	versions instead of hard-coding them.