Fix scrolling on issue notes once & for all (internal & external)...?
Widened the page and set the font size (within the "pre" tags) to be 10px, so it 
shouldn't cause scrolling even if the page is resized.  If the user has some 
feature on that forces fonts to be larger than normal... well, I can't help 
what the page looks like then.