Implemented WORKINGONIT so the site won't be accessible if that's set...
FIXES ISSUE:::
#61: WORKINGONIT isn't implemented

Apparently, I didn't check to see that this was implemented at all.  I became 
vaguely aware of the possibility when I was writing a note on issue #36, but 
dismissed it.

IMPORTANT NOTE: with this implementation, instead of just being able to shutdown 
access to the site/database, WORKINGONIT can be set to a string which can help 
in determining if something is truly broken or not: it can be manually updated 
so the string explains what's going on, or set by the upgrade system so any user 
that happens to run into it can know if there's something to worry about or not.