Fix silly problem with name of project/issue not appearing in title anymore.
In release 1.0.7, create_page_title was changed to have everything after the 
portion that checked for valid title/cache into an else statement.  Therefore 
part of the code wouldn't run, which was the part that would set the smaller 
bits of the title.  Guess I should've looked at the change a bit more carefully.