Code to update things (updates are "all-or-none"), & display attrib option list.
Handles updating, creating, & deleting attributes.  The main contact data could 
be updated, though special handling has to be implemented for the multiple email 
address situation: it should show a list of all associated email addresses, with 
some sort of indicator as to which is the primary.

The message displayed isn't very user-friendly, either.  It only shows the number 
of successes vs. failures, and is not indicative of errors (which, at present, 
cause exceptions that are handled by contentSystem::finish().

dbAbstract::db is now a public property to allow the contacts script to handle a 
transaction properly.