Retreives list of email addresses, updated template to allow display (not parsed).
Need to be able to show a static list of addresses along with the editable list... 
since there's a unique constraint in the database, there needs to be a way to 
show contacts that have the associated email, which would then allow for 
consolidation... for now, I think it can just throw an error if they try to add 
an address that's associated with another user.

On a good note: the logs table doesn't use contact_id's, so consolidating them 
is possible.  Deleting users isn't, but that is why they can be disabled.