User classes extend dbAbstract, start of fix for creating new users.
NOTE:: the problem with creating a contact right now is that the contact table 
has a not null + fkey reference to the contact_email_id, which means there has 
to be an email before there's a contact.  I'm trying to work out why adding the 
"DEFERRABLE INITIALLY DEFERRED" + the "NOT NULL" constraints won't let me add a 
record that initially has the contact_email_id null (so I have a contact_id to 
use when creating the email address)... Arg.

/lib/adminUserClass.php:
	* MAIN:::
		-- changed visibility of $db to public.
	* __construct():
		-- call parent::__construct() instead of the old PHP4 style "userClass()"
		constructor call.
	* create_user():
		-- call to create_contact() now passes email & company data.
		-- use run_sql() (courtesy dbAbstract, which userClass now extends) 
		instead of running queries the manual way.
		-- no more adding attributes for company & email.
	* add_user_to_group():
		-- use run_sql()...
	* create_contact():
		-- throw an exception if lname, fname, or email are zero-length.
		-- use run_sql()
		-- use transactions...
/lib/userClass.php:
	* MAIN:::
		-- requires & extends dbAbstract
		-- visibility of $db changed to public.
	* __construct() [FORMERLY:userClass]:
		-- renamed for the sake of dbAbstract (which has an abstract "__construct()" 
		method).