Fix problem with inability to add another email address...
When clicking the "Add Email" link, then double-clicking on the text box (so my 
browser shows a list of email addresses I've already entered), the javascript 
wasn't alerted to the change.  The result seemed to indicate that the update 
worked, even though it didn't.  Added a bunch more ways that the javascript is 
activated, along with better logic once the page gets posted.