Updated documentation to reflect ongoing changes
It was originally geared toward using Subversion, but now it's on Git.
I'm not sure what the best strategy is for making releases in Git like
they were in Subversion, so... it's a work in progress.  Hopefully the
documentation helps clarify that.  Oh, and it reflects the fact that I'm
looking for help.